LABS/CS9-LambdaMap/react_django/django-src/requirements.txt
Django==2.1
django-webpack-loader==0.6.0
gunicorn==19.9.0
passlib==1.7.1
Pillow==5.2.0
PyJWT==1.6.4
python-decouple==3.1
pytz==2018.5
whitenoise==4.0
